+++ CHANGELOG.md
@@ -6,6 +6,18 @@ CHANGELOG
 _(none)_
 
 --------------------
+## 0.3.1
+* Skip flash-based source handler with DASH sources ([#14](https://github.com/videojs/http-streaming/issues/14))
+
+## 0.3.0
+* Added additional properties to the stats object ([#10](https://github.com/videojs/http-streaming/issues/10))
+
+## 0.2.1
+* Updated the mpd-parser to fix IE11 DASH support ([#12](https://github.com/videojs/http-streaming/issues/12))
+
+## 0.2.0
+* Initial DASH Support ([#8](https://github.com/videojs/http-streaming/issues/8))
+
 ## 0.1.0
 * Initial release, based on [videojs-contrib-hls 5.12.2](https://github.com/videojs/videojs-contrib-hls)